+++ repomatic/release_prep.py
@@ -54,6 +54,7 @@ from functools import cached_property
 from pathlib import Path
 
 from .changelog import Changelog
+from .config import Config
 
 if sys.version_info >= (3, 11):
     import tomllib
@@ -72,7 +73,7 @@ class ReleasePrep:
         readme_path: Path | None = None,
         default_branch: str = "main",
     ) -> None:
-        self.changelog_path = changelog_path or Path("./changelog.md").resolve()
+        self.changelog_path = changelog_path or Path(Config.changelog_location).resolve()
         self.citation_path = citation_path or Path("./citation.cff").resolve()
         self.workflow_dir = workflow_dir or Path("./.github/workflows").resolve()
         self.readme_path = readme_path or Path("./readme.md").resolve()